Tools/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
rembg>=2.0.50
Pillow>=10.0.0
